# api/openapi.yaml
openapi: 3.1.0
info:
  title: Estate Guide API
  version: 1.0.0
  description: |
    REST API for Estate Guide. Targets GCP Cloud Run, Cloud SQL (PostgreSQL), and Cloud Storage.
    - Auth via JWT (access + refresh)
    - MSG91 for OTP (brochure downloads)
    - Cookie-based anonymous wishlist (anonId)
servers:
  - url: /api/v1
    description: Relative server (Cloud Run behind proxy)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /subscribers:
    post:
      tags: [Subscribers]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriberCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscriber' }
  /subscribers/unsubscribe:
    post:
      tags: [Subscribers]
      summary: Unsubscribe via token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '204': { description: No Content }
...
  /admin/subscribers:
    get:
      tags: [Admin]
      summary: List subscribers (admin)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/SubscriberStatus' }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Subscriber' }
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /admin/subscribers/{id}:
    delete:
      tags: [Admin]
      summary: Delete subscriber
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: No Content }
  /admin/subscribers/export:
    get:
      tags: [Admin]
      summary: Export active subscribers as CSV
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: text/csv
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
  /admin/neighborhoods:
    post:
      tags: [Admin]
      summary: Create neighborhood
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NeighborhoodCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Neighborhood' }
  /admin/neighborhoods/{id}:
    put:
      tags: [Admin]
      summary: Update neighborhood
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NeighborhoodUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Neighborhood' }
    delete:
      tags: [Admin]
      summary: Delete neighborhood
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: No Content }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
...
    SubscriberStatus:
      type: string
      enum: [active, unsubscribed]
    Subscriber:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/SubscriberStatus' }
        subscribedAt: { type: string, format: date-time }
        unsubscribedAt: { type: string, format: date-time, nullable: true }
    SubscriberCreate:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        name: { type: string, nullable: true }
...
    Neighborhood:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        city: { type: string }
        description: { type: string, nullable: true }
    NeighborhoodCreate:
      type: object
      required: [name, city]
      properties:
        name: { type: string }
        city: { type: string }
        description: { type: string, nullable: true }
    NeighborhoodUpdate:
      allOf:
        - $ref: '#/components/schemas/NeighborhoodCreate'